examples/streamlit/requirements.txt
streamlit
pandas
plotly
python-dotenv
